StatixSysUI: Default to serif typeface for lockscreen clock
Default typeface on lockscreen clock is monospace and since our
default font (GSans) did not have a monospace typeface, it always
had to fallback to the default Noto Monospace.

Setting it to serif would also imply that the clock font will change
based on the font that would be selected by the ThemePicker.

Change-Id: I3f1a5b729368584af07b5e07f3d551f88d74999a